LayoutTests/imported/w3c/web-platform-tests/dom/traversal/TreeWalker-acceptNode-filter-cross-realm-expected.txt



PASS NodeFilter is cross-realm plain object without 'acceptNode' property
PASS NodeFilter is cross-realm plain object with non-callable 'acceptNode' property
PASS NodeFilter is cross-realm plain object with revoked Proxy as 'acceptNode' property
PASS NodeFilter is cross-realm non-callable revoked Proxy
PASS NodeFilter is cross-realm callable revoked Proxy
